+++ packages/code-du-travail-modeles/src/__test__/depart.mise.retraite.banque.spec.ts
@@ -1,4 +1,5 @@
 import Engine from "publicodes";
+
 import { mergeModels } from "../internal/merger";
 import { getReferences } from "../utils/GetReferences";
 import {
@@ -32,8 +33,8 @@ describe("Préavis de retraite de la CC 2120", () => {
       "Pour un employé possédant $seniority mois d'ancienneté, son préavis devrait être $expectedResult mois",
       ({ seniority, expectedResult }) => {
         const situation = engine.setSituation({
-          "contrat salarié . convention collective": "'IDCC2120'",
           "contrat salarié . ancienneté": seniority,
+          "contrat salarié . convention collective": "'IDCC2120'",
           "contrat salarié . mise à la retraite": "non",
           "contrat salarié . travailleur handicapé": "non",
         });
@@ -64,8 +65,8 @@ describe("Préavis de retraite de la CC 2120", () => {
       "Pour un employé possédant $seniority mois d'ancienneté, son préavis devrait être $expectedResult mois",
       ({ seniority, expectedResult }) => {
         const situation = engine.setSituation({
-          "contrat salarié . convention collective": "'IDCC2120'",
           "contrat salarié . ancienneté": seniority,
+          "contrat salarié . convention collective": "'IDCC2120'",
           "contrat salarié . mise à la retraite": "oui",
           "contrat salarié . travailleur handicapé": "non",
         });